[AIRLFOW-YYY] Move task commands to separate file
