+++ airflow/bin/cli.py
@@ -21,11 +21,9 @@
 
 import argparse
 import errno
-import importlib
 import json
 import logging
 import os
-import re
 import reprlib
 import signal
 import subprocess
@@ -46,16 +44,16 @@ from tabulate import tabulate, tabulate_formats
 import airflow
 from airflow import api, jobs, settings
 from airflow.api.client import get_current_api_client
-from airflow.cli.commands import role_command, rotate_fernet_key_command, sync_perm_command, user_command
+from airflow.cli.commands import (
+    role_command, rotate_fernet_key_command, sync_perm_command, task_command, user_command,
+)
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException, AirflowWebServerTimeout
-from airflow.executors import get_default_executor
-from airflow.models import DAG, Connection, DagBag, DagModel, DagPickle, DagRun, TaskInstance, Variable
-from airflow.ti_deps.dep_context import SCHEDULER_QUEUED_DEPS, DepContext
+from airflow.models import DAG, Connection, DagBag, DagModel, DagRun, TaskInstance, Variable
 from airflow.utils import cli as cli_utils, db
+from airflow.utils.cli import get_dag, process_subdir
 from airflow.utils.dot_renderer import render_dag
-from airflow.utils.log.logging_mixin import LoggingMixin, redirect_stderr, redirect_stdout
-from airflow.utils.net import get_hostname
+from airflow.utils.log.logging_mixin import LoggingMixin
 from airflow.utils.timezone import parse as parsedate
 from airflow.www.app import cached_app, create_app
 
@@ -120,38 +118,6 @@ def setup_locations(process, pid=None, stdout=None, stderr=None, log=None):
         pid = os.path.join(settings.AIRFLOW_HOME, 'airflow-{}.pid'.format(process))
 
     return pid, stdout, stderr, log
-
-
-def process_subdir(subdir):
-    """Expands path to absolute by replacing 'DAGS_FOLDER', '~', '.', etc."""
-    if subdir:
-        subdir = subdir.replace('DAGS_FOLDER', DAGS_FOLDER)
-        subdir = os.path.abspath(os.path.expanduser(subdir))
-    return subdir
-
-
-def get_dag(args):
-    """Returns DAG of a given dag_id"""
-    dagbag = DagBag(process_subdir(args.subdir))
-    if args.dag_id not in dagbag.dags:
-        raise AirflowException(
-            'dag_id could not be found: {}. Either the dag did not exist or it failed to '
-            'parse.'.format(args.dag_id))
-    return dagbag.dags[args.dag_id]
-
-
-def get_dags(args):
-    """Returns DAG(s) matching a given regex or dag_id"""
-    if not args.dag_regex:
-        return [get_dag(args)]
-    dagbag = DagBag(process_subdir(args.subdir))
-    matched_dags = [dag for dag in dagbag.dags.values() if re.search(
-        args.dag_id, dag.dag_id)]
-    if not matched_dags:
-        raise AirflowException(
-            'dag_id could not be found with regex: {}. Either the dag did not exist '
-            'or it failed to parse.'.format(args.dag_id))
-    return matched_dags
 
 
 @cli_utils.action_logging
@@ -499,149 +465,6 @@ def dag_show(args):
         print(dot.source)
 
 
-def _run(args, dag, ti):
-    if args.local:
-        run_job = jobs.LocalTaskJob(
-            task_instance=ti,
-            mark_success=args.mark_success,
-            pickle_id=args.pickle,
-            ignore_all_deps=args.ignore_all_dependencies,
-            ignore_depends_on_past=args.ignore_depends_on_past,
-            ignore_task_deps=args.ignore_dependencies,
-            ignore_ti_state=args.force,
-            pool=args.pool)
-        run_job.run()
-    elif args.raw:
-        ti._run_raw_task(  # pylint: disable=protected-access
-            mark_success=args.mark_success,
-            job_id=args.job_id,
-            pool=args.pool,
-        )
-    else:
-        pickle_id = None
-        if args.ship_dag:
-            try:
-                # Running remotely, so pickling the DAG
-                with db.create_session() as session:
-                    pickle = DagPickle(dag)
-                    session.add(pickle)
-                    pickle_id = pickle.id
-                    # TODO: This should be written to a log
-                    print('Pickled dag {dag} as pickle_id: {pickle_id}'.format(
-                        dag=dag, pickle_id=pickle_id))
-            except Exception as e:
-                print('Could not pickle the DAG')
-                print(e)
-                raise e
-
-        executor = get_default_executor()
-        executor.start()
-        print("Sending to executor.")
-        executor.queue_task_instance(
-            ti,
-            mark_success=args.mark_success,
-            pickle_id=pickle_id,
-            ignore_all_deps=args.ignore_all_dependencies,
-            ignore_depends_on_past=args.ignore_depends_on_past,
-            ignore_task_deps=args.ignore_dependencies,
-            ignore_ti_state=args.force,
-            pool=args.pool)
-        executor.heartbeat()
-        executor.end()
-
-
-@cli_utils.action_logging
-def task_run(args, dag=None):
-    """Runs a single task instance"""
-    if dag:
-        args.dag_id = dag.dag_id
-
-    log = LoggingMixin().log
-
-    # Load custom airflow config
-    if args.cfg_path:
-        with open(args.cfg_path, 'r') as conf_file:
-            conf_dict = json.load(conf_file)
-
-        if os.path.exists(args.cfg_path):
-            os.remove(args.cfg_path)
-
-        conf.read_dict(conf_dict, source=args.cfg_path)
-        settings.configure_vars()
-
-    # IMPORTANT, have to use the NullPool, otherwise, each "run" command may leave
-    # behind multiple open sleeping connections while heartbeating, which could
-    # easily exceed the database connection limit when
-    # processing hundreds of simultaneous tasks.
-    settings.configure_orm(disable_connection_pool=True)
-
-    if not args.pickle and not dag:
-        dag = get_dag(args)
-    elif not dag:
-        with db.create_session() as session:
-            log.info('Loading pickle id %s', args.pickle)
-            dag_pickle = session.query(DagPickle).filter(DagPickle.id == args.pickle).first()
-            if not dag_pickle:
-                raise AirflowException("Who hid the pickle!? [missing pickle]")
-            dag = dag_pickle.pickle
-
-    task = dag.get_task(task_id=args.task_id)
-    ti = TaskInstance(task, args.execution_date)
-    ti.refresh_from_db()
-
-    ti.init_run_context(raw=args.raw)
-
-    hostname = get_hostname()
-    log.info("Running %s on host %s", ti, hostname)
-
-    if args.interactive:
-        _run(args, dag, ti)
-    else:
-        with redirect_stdout(ti.log, logging.INFO), redirect_stderr(ti.log, logging.WARN):
-            _run(args, dag, ti)
-    logging.shutdown()
-
-
-@cli_utils.action_logging
-def task_failed_deps(args):
-    """
-    Returns the unmet dependencies for a task instance from the perspective of the
-    scheduler (i.e. why a task instance doesn't get scheduled and then queued by the
-    scheduler, and then run by an executor).
-    >>> airflow tasks failed_deps tutorial sleep 2015-01-01
-    Task instance dependencies not met:
-    Dagrun Running: Task instance's dagrun did not exist: Unknown reason
-    Trigger Rule: Task's trigger rule 'all_success' requires all upstream tasks
-    to have succeeded, but found 1 non-success(es).
-    """
-    dag = get_dag(args)
-    task = dag.get_task(task_id=args.task_id)
-    ti = TaskInstance(task, args.execution_date)
-
-    dep_context = DepContext(deps=SCHEDULER_QUEUED_DEPS)
-    failed_deps = list(ti.get_failed_dep_statuses(dep_context=dep_context))
-    # TODO, Do we want to print or log this
-    if failed_deps:
-        print("Task instance dependencies not met:")
-        for dep in failed_deps:
-            print("{}: {}".format(dep.dep_name, dep.reason))
-    else:
-        print("Task instance dependencies are all met.")
-
-
-@cli_utils.action_logging
-def task_state(args):
-    """
-    Returns the state of a TaskInstance at the command line.
-    >>> airflow tasks state tutorial sleep 2015-01-01
-    success
-    """
-    dag = get_dag(args)
-    task = dag.get_task(task_id=args.task_id)
-    ti = TaskInstance(task, args.execution_date)
-    print(ti.current_state())
-
-
 @cli_utils.action_logging
 def dag_state(args):
     """
@@ -696,17 +519,6 @@ def dag_list_dags(args):
 
 
 @cli_utils.action_logging
-def task_list(args, dag=None):
-    """Lists the tasks within a DAG at the command line"""
-    dag = dag or get_dag(args)
-    if args.tree:
-        dag.tree_view()
-    else:
-        tasks = sorted([t.task_id for t in dag.tasks])
-        print("\n".join(sorted(tasks)))
-
-
-@cli_utils.action_logging
 def dag_list_jobs(args, dag=None):
     """Lists latest n jobs"""
     queries = []
@@ -736,82 +548,6 @@ def dag_list_jobs(args, dag=None):
                        [field.capitalize().replace('_', ' ') for field in fields],
                        tablefmt=args.output)
         print(msg)
-
-
-@cli_utils.action_logging
-def task_test(args, dag=None):
-    """Tests task for a given dag_id"""
-    # We want log outout from operators etc to show up here. Normally
-    # airflow.task would redirect to a file, but here we want it to propagate
-    # up to the normal airflow handler.
-    logging.getLogger('airflow.task').propagate = True
-
-    dag = dag or get_dag(args)
-
-    task = dag.get_task(task_id=args.task_id)
-    # Add CLI provided task_params to task.params
-    if args.task_params:
-        passed_in_params = json.loads(args.task_params)
-        task.params.update(passed_in_params)
-    ti = TaskInstance(task, args.execution_date)
-
-    try:
-        if args.dry_run:
-            ti.dry_run()
-        else:
-            ti.run(ignore_task_deps=True, ignore_ti_state=True, test_mode=True)
-    except Exception:  # pylint: disable=broad-except
-        if args.post_mortem:
-            try:
-                debugger = importlib.import_module("ipdb")
-            except ImportError:
-                debugger = importlib.import_module("pdb")
-            debugger.post_mortem()
-        else:
-            raise
-
-
-@cli_utils.action_logging
-def task_render(args):
-    """Renders and displays templated fields for a given task"""
-    dag = get_dag(args)
-    task = dag.get_task(task_id=args.task_id)
-    ti = TaskInstance(task, args.execution_date)
-    ti.render_templates()
-    for attr in task.__class__.template_fields:
-        print(textwrap.dedent("""\
-        # ----------------------------------------------------------
-        # property: {}
-        # ----------------------------------------------------------
-        {}
-        """.format(attr, getattr(task, attr))))
-
-
-@cli_utils.action_logging
-def task_clear(args):
-    """Clears all task instances or only those matched by regex for a DAG(s)"""
-    logging.basicConfig(
-        level=settings.LOGGING_LEVEL,
-        format=settings.SIMPLE_LOG_FORMAT)
-    dags = get_dags(args)
-
-    if args.task_regex:
-        for idx, dag in enumerate(dags):
-            dags[idx] = dag.sub_dag(
-                task_regex=args.task_regex,
-                include_downstream=args.downstream,
-                include_upstream=args.upstream)
-
-    DAG.clear_dags(
-        dags,
-        start_date=args.start_date,
-        end_date=args.end_date,
-        only_failed=args.only_failed,
-        only_running=args.only_running,
-        confirm_prompt=not args.yes,
-        include_subdags=not args.exclude_subdags,
-        include_parentdag=not args.exclude_parentdag,
-    )
 
 
 def get_num_ready_workers_running(gunicorn_master_proc):
@@ -2082,13 +1818,13 @@ class CLIFactory:
             'name': 'tasks',
             'subcommands': (
                 {
-                    'func': task_list,
+                    'func': task_command.task_list,
                     'name': 'list',
                     'help': "List the tasks within a DAG",
                     'args': ('dag_id', 'tree', 'subdir'),
                 },
                 {
-                    'func': task_clear,
+                    'func': task_command.task_clear,
                     'name': 'clear',
                     'help': "Clear a set of task instance, as if they never ran",
                     'args': (
@@ -2097,13 +1833,13 @@ class CLIFactory:
                         'only_running', 'exclude_subdags', 'exclude_parentdag', 'dag_regex'),
                 },
                 {
-                    'func': task_state,
+                    'func': task_command.task_state,
                     'name': 'state',
                     'help': "Get the status of a task instance",
                     'args': ('dag_id', 'task_id', 'execution_date', 'subdir'),
                 },
                 {
-                    'func': task_failed_deps,
+                    'func': task_command.task_failed_deps,
                     'name': 'failed_deps',
                     'help': (
                         "Returns the unmet dependencies for a task instance from the perspective "
@@ -2113,13 +1849,13 @@ class CLIFactory:
                     'args': ('dag_id', 'task_id', 'execution_date', 'subdir'),
                 },
                 {
-                    'func': task_render,
+                    'func': task_command.task_render,
                     'name': 'render',
                     'help': "Render a task instance's template(s)",
                     'args': ('dag_id', 'task_id', 'execution_date', 'subdir'),
                 },
                 {
-                    'func': task_run,
+                    'func': task_command.task_run,
                     'name': 'run',
                     'help': "Run a single task instance",
                     'args': (
@@ -2129,7 +1865,7 @@ class CLIFactory:
                         'ignore_depends_on_past', 'ship_dag', 'pickle', 'job_id', 'interactive',),
                 },
                 {
-                    'func': task_test,
+                    'func': task_command.task_test,
                     'name': 'test',
                     'help': (
                         "Test a task instance. This will run a task without checking for "
